+++ src/py/_pyodide/_importhook.py
@@ -130,4 +130,65 @@ def register_js_finder() -> None:
 
     global JsProxy
     JsProxy = _pyodide_core.JsProxy
+
+    for importer in sys.meta_path:
+        if isinstance(importer, JsFinder):
+            raise RuntimeError("JsFinder already registered")
+
     sys.meta_path.append(jsfinder)
+
+
+class UnvendoredStdlibFinder(MetaPathFinder):
+    """
+    A MetaPathFinder that handles unvendored and removed stdlib modules.
+
+    This class simply raises an error if a stdlib module is unvendored or removed.
+    This needs to be added to the end of sys.meta_path, so if a unvendored stdlib
+    is already loaded via pyodide.loadPackage, it can be handled by the existing finder.
+    """
+
+    def __init__(self) -> None:
+        # `test`` is not a stdlib module, but we unvendors in anyway.
+        self.stdlibs = sys.stdlib_module_names | {"test"}
+
+        # TODO: put list of unvendored stdlibs to somewhere else?
+        self.unvendored_stdlibs = {"distutils", "test", "_ssl", "_lzma"} & self.stdlibs
+
+    def find_spec(
+        self,
+        fullname: str,
+        path: Sequence[bytes | str] | None,
+        target: ModuleType | None = None,
+    ) -> ModuleSpec | None:
+        [parent, _, _] = fullname.partition(".")
+
+        if not parent or parent in sys.modules or parent not in self.stdlibs:
+            return None
+
+        if parent in self.unvendored_stdlibs:
+            raise ModuleNotFoundError(
+                f"The module '{parent}' is unvendored from the Python standard library in the Pyodide distribution, "
+                f'you can install it by calling: await micropip.install("{parent}"). '
+                "See https://pyodide.org/en/stable/usage/wasm-constraints.html for more details."
+            )
+        else:
+            raise ModuleNotFoundError(
+                f"The module '{parent}' is removed from the Python standard library in the Pyodide distribution "
+                "due to browser limitations. "
+                "See https://pyodide.org/en/stable/usage/wasm-constraints.html for more details."
+            )
+
+
+def register_unvendored_stdlib_finder() -> None:
+    """
+    A function that adds UnvendoredStdlibFinder to the end of sys.meta_path.
+
+    Note that this finder must be placed in the end of meta_paths
+    in order to prevent any unexpected side effects.
+    """
+
+    for importer in sys.meta_path:
+        if isinstance(importer, UnvendoredStdlibFinder):
+            raise RuntimeError("UnvendoredStdlibFinder already registered")
+
+    sys.meta_path.append(UnvendoredStdlibFinder())